+++ packages/terra-clinical-data-grid/CHANGELOG.md
@@ -3,6 +3,8 @@ Changelog
 
 Unreleased
 ----------
+### Changed
+* Update Copyright to include 2020
 
 2.13.0 - (December 10, 2019)
 ------------------