+++ packages/terra-clinical-data-grid/CHANGELOG.md
@@ -3,6 +3,9 @@ Changelog
 
 Unreleased
 ----------
+
+2.14.0 - (January 17, 2020)
+------------------
 ### Added
 * Added `verticalOverflowContainerRefCallback` prop to pass callback ref to `vertical-overflow-container`.
 * Added `horizontalOverflowContainerRefCallback` prop to pass callback ref to `horizontal-overflow-container`.